Update:[zustand] 전역상태 로직 store 폴더에서 파일로 분리하여 관리
